refactor added conversions
