+++ cmd/main.go
@@ -1,31 +1,38 @@
 package main
 
 import (
+	"fmt"
 	"log"
 
+	"github.com/jmoiron/sqlx"
 	"github.com/mattribution/api/pkg/http"
 	"github.com/mattribution/api/pkg/postgres"
 )
 
+const (
+	host     = "localhost"
+	port     = 5432
+	username = "postgres"
+	password = "password"
+	dbName   = "mattribution"
+)
+
 func main() {
 
+	// Make connection to DB
+	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s "+
+		"password=%s dbname=%s sslmode=disable",
+		host, port, username, password, dbName)
+	db, err := sqlx.Open("postgres", psqlInfo)
+	if err != nil {
+		panic(err)
+	}
+
 	// Setup
-	trackService, err := postgres.NewTrackService("localhost", "postgres", "password", "mattribution", 5432)
-	if err != nil {
-		log.Printf("ERROR: %v\n", err)
-	}
-	kpiService, err := postgres.NewKPIService("localhost", "postgres", "password", "mattribution", 5432)
-	if err != nil {
-		log.Printf("ERROR: %v\n", err)
-	}
-	billingEventService, err := postgres.NewBillingEventService("localhost", "postgres", "password", "mattribution", 5432)
-	if err != nil {
-		log.Printf("ERROR: %v\n", err)
-	}
-	campaignService, err := postgres.NewCampaignService("localhost", "postgres", "password", "mattribution", 5432)
-	if err != nil {
-		log.Printf("ERROR: %v\n", err)
-	}
+	trackService := postgres.TrackService{db}
+	kpiService := postgres.KPIService{db}
+	billingEventService := postgres.BillingEventService{db}
+	campaignService := postgres.CampaignService{db}
 
 	httpHandler := http.NewHandler(trackService, kpiService, billingEventService, campaignService)
 	log.Fatal(httpHandler.Serve(":3001"))
